Moved util to public namespace and stopped potential exception in move
